작은 수정

+++ include/Component/Camera.h
@@ -79,7 +79,7 @@ public:
 	virtual Ray ScreenPointToRay(float mouseX, float mouseY, int screenWidth, int screenHeight);
 
 protected:
-	f32 fovy = 45 * DEG2RAD;
+	f32 fovy = 45.0f * (f32)DEG2RAD;
 	f32 aspectRatio = 1.0f;
 	f32 zNear = 0.001f;
 	f32 zFar = 1000.0f;
